+++ packages/core/src/view/geometry/stencil/register.ts
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import StencilShapeRegistry from './StencilShapeRegistry';
+import { StencilShapeRegistry } from './StencilShapeRegistry';
 
 /**
  * Unregister all {@link StencilShape}s from {@link StencilShapeRegistry}.
@@ -24,5 +24,5 @@ import StencilShapeRegistry from './StencilShapeRegistry';
  * @since 0.18.0
  */
 export function unregisterAllStencilShapes() {
-  StencilShapeRegistry.stencils = {};
+  StencilShapeRegistry.clear();
 }