+++ packages/html/stories/Shape.stories.js
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { Graph, CylinderShape, CellRenderer } from '@maxgraph/core';
+import { Graph, CylinderShape, ShapeRegistry } from '@maxgraph/core';
 import { globalTypes, globalValues } from './shared/args.js';
 import { createGraphContainer } from './shared/configure.js';
 
@@ -92,7 +92,7 @@ const Template = ({ label, ...args }) => {
       }
     }
   }
-  CellRenderer.registerShape('box', BoxShape);
+  ShapeRegistry.add('box', BoxShape);
 
   // Creates the graph inside the DOM node.
   const graph = new Graph(container);