+++ packages/html/stories/Wires.stories.js
@@ -63,7 +63,6 @@ import {
   ConnectionConstraint,
   Point,
   CylinderShape,
-  CellRenderer,
   DomHelpers,
   Rectangle,
   EdgeHandler,
@@ -82,6 +81,7 @@ import {
   PopupMenuHandler,
   cellArrayUtils,
   StyleDefaultsConfig,
+  ShapeRegistry,
 } from '@maxgraph/core';
 
 import {
@@ -603,9 +603,7 @@ const Template = ({ label, ...args }) => {
     }
   }
 
-  CellRenderer.registerShape('resistor', ResistorShape);
-
-  // Implements a custom resistor shape. Direction currently ignored here.
+  ShapeRegistry.add('resistor', ResistorShape);
 
   const WireConnector = function (state, source, target, hints, result) {
     // Creates array of all way- and terminal points